backend/requirements.txt
# Flask and extensions
Flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0

# AI Integration
google-genai
requests>=2.31.0

# RAG — Document Knowledge Base
chromadb>=0.5.0
PyPDF2>=3.0.0
python-docx>=1.1.0

# Database
SQLAlchemy>=2.0.25
Flask-SQLAlchemy>=3.1.1

# JWT for authentication
PyJWT>=2.8.0

# Production server
gunicorn
cryptography
